DP code review comments applied.

+++ src/core/query/QueryResultsInternal.h
@@ -199,7 +199,7 @@ public:
     bool resultsApproximated;
 
     // This member keeps the estimated number of results in case of top k, if all results are actually calculated, this value is -1
-    long int estimatedNumberOfResults;
+    int estimatedNumberOfResults;
 	// map of attribute name to : "aggregation results for categories"
 	// map<string, vector<Score>>
     /*
